+++ src/main/java/roomescape/dto/request/PaymentRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 public record PaymentRequest(int amount, String paymentKey, String orderId) {
 }
