+++ src/test/java/roomescape/unit/presentation/PaymentClientControllerTest.java
@@ -19,7 +19,7 @@ import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestClient;
 import roomescape.config.paymentResponseErrorHandler;
 import roomescape.domain.PaymentInfo;
-import roomescape.dto.PaymentRequest;
+import roomescape.dto.request.PaymentRequest;
 import roomescape.exception.FilteredPaymentException;
 import roomescape.presentation.PaymentClientController;
 
@@ -28,7 +28,7 @@ public class PaymentClientControllerTest {
     private final RestClient.Builder testBuilder = RestClient.builder()
             .baseUrl("https://api.tosspayments.com")
             .defaultHeader("Authorization", String.format("%s %s", "Basic", Base64.getEncoder()));
-    
+
     private final paymentResponseErrorHandler paymentResponseErrorHandler = new paymentResponseErrorHandler();
 
     private MockRestServiceServer server = MockRestServiceServer.bindTo(testBuilder).build();
